Reorganize test/internal: build-system tests to source-lints/, delete redundant tests
Move build-system tests (oxlint-plugin-bun, expected-durations,
build-post-link-ordering, macos-cross-config, rust-windows-sys-link) into
test/internal/source-lints/ so they run once on GitHub Actions instead of
on every Buildkite lane. Tests of bun-executable internals (bindgen, fifo,
highlighter, powershell-escape, internal-module-blob, features-json) stay
in test/internal/ and continue to run on Buildkite against the PR build.

Delete three redundant tests and their internal-for-testing probes:
  - linear-fifo.test.ts: same cases covered by #[cfg(test)] in
    src/collections/linear_fifo.rs
  - sigaction-layout.test.ts: Zig-era concern; Rust uses the libc crate's
    per-target sigaction layout
  - int_from_float.test.ts: tested a Zig-only helper; the CSS crash case
    it guarded is already in test/js/bun/css/css.test.ts

Split the features.json parser vs binary check out of macos-cross-config
into its own test/internal/features-json.test.ts since that one needs the
PR binary.

Add the placement rule to REVIEW.md and update the source-lints README.
