+++ src/js/internal-for-testing.ts
@@ -306,11 +306,6 @@ export const decodeURIComponentSIMD = $newCppFunction(
 
 export const getDevServerDeinitCount = $bindgenFn("DevServer.bind.ts", "getDeinitCountForTesting");
 export const getCounters = $newRustFunction("Counters.rs", "createCountersObject", 0);
-export const linearFifoOrderedRemoveProbe = $newRustFunction(
-  "collections/linear_fifo.rs",
-  "TestingAPIs.orderedRemoveProbe",
-  1,
-) as (scenario: number) => number[];
 export const hasNonReifiedStatic = $newCppFunction("InternalForTesting.cpp", "jsFunction_hasReifiedStatic", 1);
 
 interface setSocketOptionsFn {
@@ -446,14 +441,6 @@ export const sysErrorNameFromLibuv: (errno: number) => string | undefined = $new
   "TestingAPIs.sysErrorNameFromLibuv",
   1,
 );
-
-export const sigactionLayout: () =>
-  | undefined
-  | {
-      installed: { handler: number; flags: number };
-      readback: { handler: number; flags: number };
-      sizeof: number;
-    } = $newRustFunction("sys.rs", "TestingAPIs.sigactionLayout", 0);
 
 export const stringsInternals = {
   /**
