+++ test/internal/features-json.test.ts
@@ -1,0 +1,34 @@
+// Cross lanes can't run the binary they just linked, so features.json is
+// generated by parsing PACKED_FEATURES_LIST out of src/analytics/lib.rs
+// instead of asking the binary. The list maps crash-report bit indices to
+// feature names and order matters. This pins the parser against what the
+// binary built from the same source tree reports, so a macro reshape that
+// breaks the parser fails here instead of silently shipping a misaligned
+// feature table.
+import { expect, test } from "bun:test";
+import { bunEnv, bunExe } from "harness";
+import { join } from "node:path";
+
+import { parsePackedFeaturesList } from "../../scripts/build/features-json.ts";
+
+test("parsePackedFeaturesList matches crash_handler.getFeatureData() exactly", async () => {
+  const repoRoot = join(import.meta.dir, "..", "..");
+  const parsed = parsePackedFeaturesList(repoRoot);
+  expect(parsed.length).toBeGreaterThan(0);
+
+  await using proc = Bun.spawn({
+    cmd: [
+      bunExe(),
+      "-e",
+      `const { crash_handler } = require("bun:internal-for-testing"); console.log(JSON.stringify(crash_handler.getFeatureData().features));`,
+    ],
+    env: { ...bunEnv, BUN_FEATURE_FLAG_INTERNAL_FOR_TESTING: "1" },
+    stderr: "pipe",
+  });
+  const [stdout, stderr, exitCode] = await Promise.all([proc.stdout.text(), proc.stderr.text(), proc.exited]);
+  if (exitCode !== 0) {
+    expect(stderr).toBe("");
+  }
+  expect(exitCode).toBe(0);
+  expect(parsed).toEqual(JSON.parse(stdout));
+});
