+++ test/internal/source-lints/macos-cross-config.test.ts
@@ -8,21 +8,20 @@
  * where the same inputs intentionally resolve to the native toolchain instead.
  */
 import { describe, expect, test } from "bun:test";
-import { bunEnv, bunExe, isMacOS, tempDir } from "harness";
+import { isMacOS, tempDir } from "harness";
 import { existsSync, mkdirSync, writeFileSync } from "node:fs";
 import { join } from "node:path";
 
-import { resolveConfig, type Config, type PartialConfig, type Toolchain } from "../../scripts/build/config.ts";
-import { webkit } from "../../scripts/build/deps/webkit.ts";
-import { parsePackedFeaturesList } from "../../scripts/build/features-json.ts";
-import { computeFlags, DARWIN_STACK_SIZE } from "../../scripts/build/flags.ts";
-import { MACOS_SDK_VERSION, macosSdkCachePath, resolveMacosSdkPath } from "../../scripts/build/macos-sdk.ts";
-import { rustTarget } from "../../scripts/build/rust.ts";
+import { resolveConfig, type Config, type PartialConfig, type Toolchain } from "../../../scripts/build/config.ts";
+import { webkit } from "../../../scripts/build/deps/webkit.ts";
+import { computeFlags, DARWIN_STACK_SIZE } from "../../../scripts/build/flags.ts";
+import { MACOS_SDK_VERSION, macosSdkCachePath, resolveMacosSdkPath } from "../../../scripts/build/macos-sdk.ts";
+import { rustTarget } from "../../../scripts/build/rust.ts";
 import {
   elfDebugCompressPostlinkCommand,
   machoEntitlementsPlist,
   machoPostlinkCommand,
-} from "../../scripts/build/shims.ts";
+} from "../../../scripts/build/shims.ts";
 
 /** A fully-populated fake toolchain — resolveConfig never spawns any of these. */
 function mockToolchain(overrides: Partial<Toolchain> = {}): Toolchain {
@@ -254,37 +253,6 @@ describe.skipIf(isMacOS)("macOS cross-compile config (non-darwin host)", () => {
     expect(flags.cxxflags.some(f => f.includes("apple-macosx"))).toBe(false);
     expect(flags.cxxflags).not.toContain("-isysroot");
     expect(flags.cxxflags).not.toContain("-nostdinc");
-  });
-});
-
-describe("host-side features.json for cross-compiled binaries", () => {
-  // Cross lanes can't run the binary they just linked, so features.json is
-  // generated by parsing PACKED_FEATURES_LIST out of src/analytics/lib.rs
-  // instead of asking the binary. The list maps crash-report bit indices to
-  // feature names — order matters. This pins the parser against what the
-  // *actual* binary built from the same source tree reports, so a macro
-  // reshape that breaks the parser fails here instead of silently shipping
-  // a misaligned feature table.
-  test("parsed feature list matches crash_handler.getFeatureData() exactly", async () => {
-    const repoRoot = join(import.meta.dir, "..", "..");
-    const parsed = parsePackedFeaturesList(repoRoot);
-    expect(parsed.length).toBeGreaterThan(0);
-
-    await using proc = Bun.spawn({
-      cmd: [
-        bunExe(),
-        "-e",
-        `const { crash_handler } = require("bun:internal-for-testing"); console.log(JSON.stringify(crash_handler.getFeatureData().features));`,
-      ],
-      env: { ...bunEnv, BUN_FEATURE_FLAG_INTERNAL_FOR_TESTING: "1" },
-      stderr: "pipe",
-    });
-    const [stdout, stderr, exitCode] = await Promise.all([proc.stdout.text(), proc.stderr.text(), proc.exited]);
-    if (exitCode !== 0) {
-      expect(stderr).toBe("");
-    }
-    expect(exitCode).toBe(0);
-    expect(parsed).toEqual(JSON.parse(stdout));
   });
 });
